+++ src/Core/Models/StaticStore/Plan.cs
@@ -17,7 +17,7 @@ namespace Bit.Core.Models.StaticStore
         public short? MaxUsers { get; set; }
 
         public bool HasAdditionalSeatsOption { get; set; }
-        public short? MaxAdditionalSeats { get; set; }
+        public int? MaxAdditionalSeats { get; set; }
         public bool HasAdditionalStorageOption { get; set; }
         public short? MaxAdditionalStorage { get; set; }
         public bool HasPremiumAccessOption { get; set; }
